Streamlining ADL library to match CCD
Doing initial work to strip out the older
'buffered' ADL library that I'd extended from
the official AMD sample, in order to make it
closer to the simple, small one I made for the
Windows CCD interface. The aim is to slim
down the codebase and to use some of the
struct tips I learned to simplify accessing the
flags that AMD uses for it's settings.