Fleshed out known bitvectors
The AMD documentation is woeful! There are so
many missing bitvector values and the relationship
between those values and the Mask and Value fields
are only intermittently described! It's pretty crap.

I think I'm going to need to ask the developers some
questions in order to get the required values.