[WIP] Trying to get ADL2_Display_SLSMapConfigX2_Get working
This should be working. It matches other examples I can find, I'm
calling the function with the right structures and the right order
as far as the documentation states, but it is always returning ADL_ERR.

It's odd because I know that we have a working SLS grid, and the
ADL2_Display_SLSMapIndex_Get is returning the SLS Map index (4),
and yet when the exact same adapter index, slsimap index and context
handle are passed to ADL2_Display_SLSMapConfigX2_Get it errors!

EIther the documentation is incorrect, and that is causing a problem,
or there is actually a problem with the ADL2_Display_SLSMapConfigX2_Get
function. I tried the ADL2_Display_SLSMapConfig_Get as well, and the same
problem.