[WIP] works to set and unset Eyefinity
Does some weird multiple attempts at settings
Eyefinity, which I have to figure out how to stop.
I now don't think that the ADL_Display_SLSMapConfigX2_Get
function actually works. But luckily it's not needed
in order to turn Eyefinity on and off (which is what we
want to do).

Next step is to figure out how to reduce the data I record
to the minimum needed, and then add AMD HDR settings.