[WIP] Slightly different ADL restructure
Made changes to the AMDLibrary so that it
follows the structure of the Eyefinity sample
eyefinity.cpp provided within the AMD ADL
library. Am trying to grab the details of the
adapters and displays in order to properly call
the SLS get function.

Also added a DLL check to make sure that if
the AMD ADL DLL doesn't exist then the code
handles it by completing, but leaves the
AMDLibrary.IsInstalled as false.