[WIP] Adding missing functions to AMDLIbrary
This copies across some missing functions to AMDLibrary from NVIDIALibrary. It hasn't made any changes to those functions though ... yet. I will be starting that journey over the next couple of weeks. At this stage it's just making sure that the AMDLibrary sturctures and API matches that of the NVIDIALibrary as that one is working. The thinking is that if this one is built in the same structure as the NVIDIALibrary one, then it's more likely to work, and will be much easier to troubleshoot.