Fixed broken SLS detection
SLS detection logic was broken, which then mean that the library was trying to set the AMD SLS function when it shouldn't have been doing so. Additionally, the HDR CONFIG WAS MISSING THE ADAPTER INDEX, WHICH MEANT WE WERE TRYING TO SET THE hdr MULTIPLE TIMES INSTEAD OF JUST DOING IT ONCE. Oops caps lock!

Also Changed the AMDInfo program version as I'm going to do a new release once I've solved the current issues.