Fixed SLS detection logic
So fixed the SLS detection logic during the configuration saving process. AMD have made it unbelievably complicated to find out whether Eyefinity is enabled, and it's such a stupid thing to have to do. It could be so much easier if they developed a ADL2_Display_SLSMapConfig_GetState  command that you could pass a simple SLSMap and it would return an ADL_TRUE or ADL_FALSE.

One thing that needs correcting is that we currently record the SLS configuration when SLS is not enabled. This wastes space in the display configuration, which will ebcome an issue when this is eventually integrated into DisplayMagician. Fixing that is my next challenge.