Added create SLS map if not exists in driver
Due to the AMD Radeon UI being so darn confusing for Eyefinity, I've now added the ability for this library to recognise when an Eyefinity configuration has been accidentally deleted by a user from the AMD Radeon UI, and it will attempt to create the Eyefinity config again itself, and turn it on. This avoids any issues that the AMD Radeon UI causes, and will save me lots of support calls!