Added a WinLibrary update config to refresh config
I *think* that AMD Eyefinity changes are so great that WinLibrary gets confused and gives up. AMD Eyefinity affects the windows display layout so much. This change adds a windows config refresh after the AMD config is applied to make sure that the winLibrary calls are using the latest updated active config that results from the AMDLibrary config changes.