Updated WinLibrary to latest version
This bugfix means all adapters known to windows are stored in each profile, which means that if any video cards are added or removed, then the display profiles won't match, even if those adapters aren't used in the display layout. This change was made to ensure that all cloned displays will work. This change makes DisplayMagician more inline with our goal - if the hardware changes then displaymagician should require the user to create new display profiles.